feat: Implement rounded background text widget and field
- Added `LineMetricsModel` to encapsulate text line metrics with support for custom widths and corner rounding.
- Created `RoundedBackgroundTextPainter` for rendering text with a rounded background and cursor support.
- Introduced `RoundedBackgroundText` widget for displaying text with optional styling and background.
- Developed `RoundedBackgroundTextField` for editable text with a rounded background, supporting various customization options.
- Updated import paths for better organization and accessibility of the new features.
